# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hungry-Hub Food Delivery API
  description: Minimal OpenAPI spec covering auth, restaurants, menu, and orders (initial draft)
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      properties:
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        coordinates:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
    Restaurant:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        address: { $ref: '#/components/schemas/Address' }
        deliveryFee: { type: number }
        isOpen: { type: boolean }
...
paths:
...
  /api/v1/restaurants:
    get:
      summary: List restaurants (supports query filters)
      parameters:
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: cuisine
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
  /api/v1/restaurants/{id}:
    get:
      summary: Get restaurant details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
...
security:
  - bearerAuth: []